Support --local and --quickstart testing
Controlled by the (new and optional) "type" field in the instance
configuration, which would be "local" or "quickstart".  Such an instance
bypasses the "max-instances" setting (IOW, can run in parallel with any
number of other instances.)  Snapshot selection is also bypassed, since
there is no VM to restore a snapshot in.

If --local fails due to there being no tests, it reports "not supported"
instead of failing.