Use the --vm-web-server testing framework argument if available
The argument selects which web server configuration to install (one of
Apache+mod_wsgi, nginx+uWSGI or "pure" uWSGI.)

Also add a workaround for Ubuntu 14.04 LTS where we will test the default
configuration (nginx+uWSGI) once supported, and thus have that installed
in the 'with-prereqs' snapshot.  Older commits will still install the
Apache+mod_wsgi configuration, and that won't work if nginx is already
installed, so force use of the 'clean' snapshot then.