Add support for multiple required Python versions
The correct one for a given commit is detected by loading and executing
the `pythonversion.py` script in that commit with different Python
interpreters.

A new configuration setting, "python-versions" define available versions
and paths to executables. The executable paths need to be the same on
the system running this script as on the VM.