Add support for commits declaring that testing is broken
This is declared by flags under testing/flags/. If `broken-all.flag`
exists, all testing is disabled. If `broken-$identifier.flag` exists,
testing in the specified type of instance is disabled.

Broken testing is not considered an error; it's treated as commits not
supporting a certain type of instancen / OS version. The purpose is to
enable branches that do major changes split into multiple commits,
where intermediate commits make sense in terms of commit history, but
inevitably break testing.

This could also be used as a more convenient way to disable an older OS
version once support for it is dropped, but still test older branches
on that OS version.